Fix SPIRE block rank assignment flag decoding

+++ crates/ecaz-cli/src/commands/bench/spire_pipeline.rs
@@ -1466,7 +1466,7 @@ struct LeafBlockRankRow {
     row_end: Option<i64>,
     row_count: Option<i64>,
     block_ip: Option<f32>,
-    assignment_flags: Option<i64>,
+    assignment_flags: Option<i32>,
 }
 
 impl From<Row> for LeafBlockRankRow {
@@ -1518,7 +1518,7 @@ struct LeafBlockRankRecord {
     row_end: Option<i64>,
     row_count: Option<i64>,
     block_ip: Option<f32>,
-    assignment_flags: Option<i64>,
+    assignment_flags: Option<i32>,
 }
 
 impl LeafBlockRankRecord {
